test run with no obs error or surv selex
EMs are getting the catch comps fairly well which makes me think that the issues that are muddying the survey are leading to the failed fits.
So we are going to try a run with 100% survey coverage, no obs error, and perfect 100% selex to see where the scaling issue arises.