fix comparison values
for reasons that are still somewhat unclear, it looks like the total biomass coming out of WHAM (which I have to back-calculate by ignoring maturity) is roughly equal to the putative "true" ssb (penalized by maturity) in the model. this is true even when I correct the timing for the SSB calculation. Will need to dig into this after my talk.